[Rename 🏷️] customTagProps.ts -> customTag.ts로 변경
